move algolia client to hook
